Divided server.cpp into 2 new srcs files launch.cpp and monitor.cpp
